+++ src/App.vue
@@ -46,6 +46,8 @@ import MainView from '@/views/MainView.vue';
 import {useStore} from '@/stores/main';
 import {useConfirmDialog} from '@/vue-plugins/plugin-confirm-dialog';
 import printToHtml from '@/print/print';
+import type {PrintTranslations} from '@/print/print';
+import {unflatten} from 'flat';
 
 const store = useStore();
 const {data, dirty} = storeToRefs(store);
@@ -70,6 +72,7 @@ const mdlSave = ref<InstanceType<typeof ModalChartSave>>();
 
 onMounted(() => {
   resetModel();
+  setPageTitle();
 });
 
 const setPageTitle = () => {
@@ -100,22 +103,29 @@ const onChartNew = async () => {
 };
 
 const resetModel = () => {
-  const filename = t('misc.default-filename');
-  store.newModel(filename);
+  store.newModel();
 };
 
 const onChartSave = () => {
   mdlSave.value?.show();
 };
 const onChartSaveAs = (filename: string) => {
-  const doc = printToHtml(data.value);
+  const translations = unflatten<{[key: string]: string}, PrintTranslations>(
+    Object.fromEntries(
+      store.saveTemplateKeys.map((k) => {
+        const tkey = k.replace(/^t\./, '');
+        return [tkey, t(`save-template.${tkey}`)];
+      }),
+    ),
+  );
+  const doc = printToHtml(data.value, translations);
   const blob = new Blob([doc], {type: 'text/html;charset=utf-8'});
   saveAs(blob, filename);
   dirty.value = false;
 };
 
-const onChartOpen = (hmtlSource: string) => {
-  store.loadHtmlAsModel(hmtlSource);
+const onChartOpen = (hmtlSource: string, filename: string) => {
+  store.loadHtmlAsModel(hmtlSource, filename);
 };
 
 const beforeUnloadHandler = (evt: Event) => {